+++ common/functions/tests/it/scalars/strings/trim.rs
@@ -14,24 +14,15 @@
 
 use common_datavalues::prelude::*;
 use common_exception::Result;
-use common_functions::scalars::Function;
 use common_functions::scalars::LTrimFunction;
 use common_functions::scalars::RTrimFunction;
 use common_functions::scalars::TrimFunction;
-use pretty_assertions::assert_eq;
+
+use super::Test;
+use super::run_tests;
 
 #[test]
 fn test_trim_function() -> Result<()> {
-    struct Test {
-        display: &'static str,
-        nullable: bool,
-        arg_names: Vec<&'static str>,
-        columns: Vec<DataColumn>,
-        expect: DataColumn,
-        error: &'static str,
-        func: Box<dyn Function>,
-    }
-
     let schema = DataSchemaRefExt::create(vec![
         DataField::new("a", DataType::String, false),
         DataField::new("b", DataType::Int64, false),
@@ -40,6 +31,7 @@ fn test_trim_function() -> Result<()> {
 
     let tests = vec![
         Test {
+            name: "ltrim-abc-passed",
             display: "ltrim",
             nullable: true,
             arg_names: vec!["a"],
@@ -49,6 +41,7 @@ fn test_trim_function() -> Result<()> {
             error: "",
         },
         Test {
+            name: "rtrim-abc-passed",
             display: "rtrim",
             nullable: true,
             arg_names: vec!["a"],
@@ -58,6 +51,7 @@ fn test_trim_function() -> Result<()> {
             error: "",
         },
         Test {
+            name: "trim-abc-passed",
             display: "trim",
             nullable: true,
             arg_names: vec!["a"],
@@ -67,6 +61,7 @@ fn test_trim_function() -> Result<()> {
             error: "",
         },
         Test {
+            name: "trim-blank-passed",
             display: "trim",
             nullable: true,
             arg_names: vec!["a"],
@@ -76,48 +71,5 @@ fn test_trim_function() -> Result<()> {
             error: "",
         },
     ];
-
-    for t in tests {
-        let func = t.func;
-        let rows = t.columns.len();
-
-        // Type check.
-        let mut args = vec![];
-        let mut fields = vec![];
-        for name in t.arg_names {
-            args.push(schema.field_with_name(name)?.data_type().clone());
-            fields.push(schema.field_with_name(name)?.clone());
-        }
-
-        let columns: Vec<DataColumnWithField> = t
-            .columns
-            .iter()
-            .zip(fields.iter())
-            .map(|(c, f)| DataColumnWithField::new(c.clone(), f.clone()))
-            .collect();
-
-        if let Err(e) = func.eval(&columns, rows) {
-            assert_eq!(t.error, e.to_string());
-        }
-        func.eval(&columns, rows)?;
-
-        // Display check.
-        let expect_display = t.display.to_string();
-        let actual_display = format!("{}", func);
-        assert_eq!(expect_display, actual_display);
-
-        // Nullable check.
-        let expect_null = t.nullable;
-        let actual_null = func.nullable(&schema)?;
-        assert_eq!(expect_null, actual_null);
-
-        let v = &(func.eval(&columns, rows)?);
-
-        // Type check.
-        let expect_type = func.return_type(&args)?;
-        let actual_type = v.data_type();
-        assert_eq!(expect_type, actual_type);
-        assert_eq!(v, &t.expect);
-    }
-    Ok(())
+    run_tests(tests, schema)
 }
